master thread introduced
We' d like to arrange the threads so they works sequentially by the
order of terms, however threads are awaken stochastically. Then the
"master" thread has to be suspended to wait for the thread which
processes the front terms being finished, resulting the same fact that
too many termporary files generated in the directory. I give up, and
manually set the checkpoint, when any the background thread reaches the
checkpoint they are suspended to be synchronized and if all threads are
synchronized, the merge(master) thread is awaken, then go on to the next
checkpoint.
it is reached the merge thread is awaken, all the other threads stops
until merge is done.